Add test to verify df -h output after replace, expand and shrink ops
Test to verify the df -h output when for a given volume, the bricks are
replaces, volume size is shrinked and when the volume size is expanded.
Steps:
- Take the output of df -h.
- Replace any one brick for the volumes.
- Wait till the heal is completed
- Repeat steps 1, 2 and 3 for all bricks for all volumes.
- Check if there are any inconsistencies in the output of df -h
- Remove bricks from volume and check output of df -h
- Add bricks to volume and check output of df -h

The size of mount points should remain unchanged during a replace op,
and the sizes should vary according to shrink or expand op performed
on the volume.

Change-Id: I323da4938767cad1976463c2aefb6c41f355ac57
Signed-off-by: Pranav <prprakas@redhat.com>